cue-backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.3,<6.0
celery>=5.4,<6.0
redis>=5.0,<6.0
psycopg[binary]>=3.2,<4.0
django-cors-headers>=4.6,<5.0
python-dotenv>=1.0,<2.0
openai>=1.60,<2.0
google-auth>=2.35,<3.0
